requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
python-dotenv==1.0.1
Werkzeug==3.0.4

